% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{cars_mlp_itr}
\alias{cars_mlp_itr}
\title{Iterative optimization of neural network}
\value{
An object with primary class \code{iteration_results}.
}
\description{
This object has the results when a neural network was tuned using Bayesian
optimization and 10-fold cross-validation.
}
\details{
The code used to produce this object:

library(tidymodels)
tidymodels_prefer()
}
\section{------------------------------------------------------------------------------}{
set.seed(1)
cars_rs <- vfold_cv(mtcars)

car_rec <-
recipe(mpg ~ ., data = mtcars) \%>\%
step_normalize(all_predictors())

mlp_spec <-
mlp(hidden_units = tune(),
penalty = tune(),
epochs = tune()) \%>\%
set_mode("regression")

set.seed(1)
cars_mlp_itr <-
mlp_spec \%>\%
tune_bayes(
car_rec,
resamples = cars_rs,
initial = 5,
iter = 4,
control = control_bayes(save_pred = TRUE)
)\preformatted{}
}

\keyword{datasets}
