% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\alias{explore.default}
\alias{explore.tune_results}
\title{Launches shiny application based on data from objects}
\usage{
explore(x, hover_cols = NULL, ...)

\method{explore}{default}(x, hover_cols = NULL, ...)

\method{explore}{tune_results}(x, hover_cols = NULL, hover_only, ...)
}
\arguments{
\item{x}{The \code{\link[tune:fit_resamples]{tune::fit_resamples()}} result.}

\item{hover_cols}{The columns to display while hovering in the Shiny app.}

\item{...}{Other parameters not currently used.}
}
\description{
This function takes the data based on \code{\link[tune:fit_resamples]{tune::fit_resamples()}} results to
launch a Shiny app. This function normally does not return; interrupt R
to stop the app (usually by pressing Ctrl+C or Esc).
}
\keyword{classes,}
\keyword{classif}
\keyword{graphs,}
\keyword{models,}
\keyword{regression,}
