% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{first_class_prob_name}
\alias{first_class_prob_name}
\title{Rwturns the name of predictions column for the first level variable}
\usage{
first_class_prob_name(predictions_obj, event_level, y_name)
}
\arguments{
\item{event_level}{give the original event level for the dataset}

\item{y_name}{give the response variable for the dataset}

\item{predictions_object}{give the predictions object from tidymodels}
}
\value{
first_class_prob_name ()
}
\description{
This function takes the predictions object, event_level, and y_name as
arguments and returns the predictions column for the first level variable.
}
\keyword{classes,}
\keyword{classif}
\keyword{models,}
