% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{format_hover}
\alias{format_hover}
\title{Returns the hover columns to be displayed in interactive plots}
\usage{
format_hover(x, ...)
}
\arguments{
\item{x}{A data frame with columns to be displayed in the hover.}

\item{...}{Other parameters not currently used.}
}
\value{
A character vector.
}
\description{
This function takes .hover argument and returns the output that can
be used as a test aesthetics in a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object to customize tooltip.
}
\keyword{models}
