% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Internal_data}
\alias{Internal_data}
\alias{ad_data_boost_res}
\alias{mtcars_spline_res}
\alias{ames_rf_res}
\alias{volcano_res}
\title{fit_resamples() object for internal testing}
\value{
a fit_resample() object
}
\description{
fit_resamples() object for internal testing
}
\details{
An internal object used for development containing the
fit_resamples() result for different models
}
\keyword{datasets}
