% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\alias{launch.default}
\alias{launch.reg_shiny_data}
\alias{launch.two_cls_shiny_data}
\title{Internal function to run shiny application on an object of \code{shiny_data} class}
\usage{
launch(x, hover_cols = NULL, hover_only = NULL, ...)

\method{launch}{default}(x, hover_cols = NULL, hover_only = NULL, ...)

\method{launch}{reg_shiny_data}(x, hover_cols = NULL, hover_only = NULL, ...)

\method{launch}{two_cls_shiny_data}(x, hover_cols = NULL, hover_only = NULL, ...)
}
\arguments{
\item{x}{The \code{\link[=organize_data]{organize_data()}} result.}

\item{hover_cols}{The columns to display while hovering in the Shiny app.}

\item{hover_only}{A Boolean to turn on and off hovering in the interactive
plots}

\item{...}{Other parameters not currently used.}
}
\description{
This function takes the \code{\link[=organize_data]{organize_data()}} result to launch a Shiny app.
}
