% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_class_diag_plots.R
\name{plot_multiclass_obs_pred}
\alias{plot_multiclass_obs_pred}
\title{Visualizing predicted probability vs. true class for a multi-class classification
model}
\usage{
plot_multiclass_obs_pred(dat, y_name, prob_bins = 0.05)
}
\arguments{
\item{dat}{The predictions data frame in the \code{\link[=organize_data]{organize_data()}} result. Following
variables are required: \code{.outcome}, \code{.pred}, \code{.color}, and \code{.hover}.}

\item{y_name}{The y/response variable for the model.}

\item{prob_bins}{The desired \code{binwidth} for histogram.}
}
\value{
A \code{\link[plotly:ggplotly]{plotly::ggplotly()}} object.
}
\description{
This function plots the predicted probabilities against the observed class based on
tidymodels results for a multi-class classification model.
}
\keyword{internal}
