% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_class_diag_plots.R
\name{plot_multiclass_roc}
\alias{plot_multiclass_roc}
\title{Visualizing the ROC curve for a classification model}
\usage{
plot_multiclass_roc(dat, y_name)
}
\arguments{
\item{dat}{The predictions data frame in the \code{\link[=organize_data]{organize_data()}} result. Following
variables are required: \code{.outcome}, \code{.pred}, \code{.color}, and \code{.hover}.}

\item{y_name}{The y/response variable for the model.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
This function plots the ROC curve for a classification model.
}
\keyword{classes,}
\keyword{classif,}
\keyword{graphs}
\keyword{models,}
