% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diag_plots.R
\name{plot_numeric_res_factorcol}
\alias{plot_numeric_res_factorcol}
\title{Visualizing residuals vs. a factor column for a regression model}
\usage{
plot_numeric_res_factorcol(dat, factorcol, y_name)
}
\arguments{
\item{dat}{give the data frame that has the original data and the results of
collect_predictions()}

\item{factorcol}{give the factor column you want to plot against the residuals}

\item{y_name}{give the y/response variable for the model}
}
\value{
plot_numeric_res_factorcol()
}
\description{
This function allows you to plot the residuals based on tidymodels result for
a regression model against any of the factor column used for modeling.
}
\keyword{graphs}
\keyword{models,}
\keyword{regression,}
