% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diag_plots.R
\name{plot_numeric_res_pred}
\alias{plot_numeric_res_pred}
\title{Visualizing residuals vs. predicted values for a model}
\usage{
plot_numeric_res_pred(original_data, object)
}
\arguments{
\item{object}{give the object you got using tidymodels::fit_resamples}
}
\value{
plot_numeric_res_pred()
}
\description{
This function allows you to plot the predicted values based on your tidymodels
result against the residuals.
}
\keyword{graphs}
\keyword{models,}
\keyword{regression,}
