% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{plot_tuning_params}
\alias{plot_tuning_params}
\title{Autoplot of tuning parameters}
\usage{
plot_tuning_params(obj)
}
\arguments{
\item{obj}{The results from tune_() functions.}
}
\value{
A \code{\link[plotly:ggplotly]{plotly::ggplotly()}} object.
}
\description{
This function plots the tuning parameters against a metric.
}
