% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_diag_plots.R
\name{plot_twoclass_conf_mat}
\alias{plot_twoclass_conf_mat}
\title{Visualizing the confusion matrix for a model}
\usage{
plot_twoclass_conf_mat(object)
}
\arguments{
\item{object}{give the object you got using tidymodels::fit_resamples}
}
\value{
plot_twoclass_obs_pred()
}
\description{
This function allows you to plot the confusion matrix for your classification
model
}
\keyword{classes,}
\keyword{classif,}
\keyword{graphs}
\keyword{models,}
