% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_diag_plots.R
\name{plot_twoclass_pr}
\alias{plot_twoclass_pr}
\title{Visualizing the PR curve for a classification model}
\usage{
plot_twoclass_pr(object, event_level = "first")
}
\arguments{
\item{object}{The object you got using \code{\link[tune:fit_resamples]{tune::fit_resamples()}}.}

\item{event_level}{The desired event level for the dataset.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
This function plots the full precision recall curve.
}
\keyword{classes,}
\keyword{classif,}
\keyword{graphs}
\keyword{models,}
