% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_diag_plots.R
\name{plot_twoclass_pred_numcol}
\alias{plot_twoclass_pred_numcol}
\title{Visualizing the predicted probability vs. a numeric variable for a model}
\usage{
plot_twoclass_pred_numcol(
  object,
  numcol,
  event_level = "first",
  prob_breaks = (2:9)/10,
  prob_eps = 0.001
)
}
\arguments{
\item{object}{give the object you got using tidymodels::fit_resamples}

\item{numcol}{give the numerical column you want to plot}
}
\value{
plot_twoclass_pred_numcol()
}
\description{
This function allows you to plot the predicted probabilities from your tidymodel
result against a numeric variable.
}
\keyword{classes,}
\keyword{classif,}
\keyword{graphs}
\keyword{models,}
