% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_diag_plots.R
\name{plot_twoclass_roc}
\alias{plot_twoclass_roc}
\title{Visualizing the roc curve for a classification model}
\usage{
plot_twoclass_roc(object, event_level = "first")
}
\arguments{
\item{object}{give the object you got using tidymodels::fit_resamples}
}
\value{
plot_twoclass_roc()
}
\description{
This function allows you to plot the roc curve for your classification model.
}
\keyword{classes,}
\keyword{classif,}
\keyword{graphs}
\keyword{models,}
