% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_organize_data.R
\name{print.shiny_data}
\alias{print.shiny_data}
\title{Returns the class, app_type, y_name, and nrows of an object of shiny_data class}
\usage{
\method{print}{shiny_data}(x, ...)
}
\arguments{
\item{x}{an object of class shiny_data}
}
\description{
This is a print method for a shiny_data class
}
