% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{explore}
\title{Objects exported from other packages}
\details{
The default configuration is based on \strong{the optimal value} of the first metric.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{explore}}}
}}

