% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymodels-package.R
\docType{package}
\name{shinymodels-package}
\alias{shinymodels}
\alias{shinymodels-package}
\title{shinymodels: Interactive Assessments of Models}
\description{
Launch a 'shiny' application for 'tidymodels' results. For classification or regression models, the app can be used to determine if there is lack of fit or poorly predicted points.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shinymodels.tidymodels.org}
  \item \url{https://github.com/tidymodels/shinymodels}
  \item Report bugs at \url{https://github.com/tidymodels/shinymodels/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Authors:
\itemize{
  \item Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})
  \item Shisham Adhikari \email{shadhikari@ucdavis.edu}
  \item Julia Silge \email{julia.silge@posit.co} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
