% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymodels.R
\name{shinymodels}
\alias{shinymodels}
\alias{shinymodels.default}
\alias{shinymodels.reg_shiny_data}
\alias{shinymodels.two_cls_shiny_data}
\title{Internal function to run shiny application on an object of \code{shiny_data} class}
\usage{
shinymodels(x, hover_only = NULL, ...)

\method{shinymodels}{default}(x, hover_cols = NULL, ...)

\method{shinymodels}{reg_shiny_data}(x, hover_only, ...)

\method{shinymodels}{two_cls_shiny_data}(x, hover_only, ...)
}
\arguments{
\item{x}{The \code{\link[=organize_data]{organize_data()}} result.}

\item{hover_only}{The Boolean to turn on and off the hover instead of selected}

\item{...}{Other parameters not currently used.}
}
\description{
This function takes the \code{\link[=organize_data]{organize_data()}} result to launch a Shiny app.
}
