% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{two_class_final}
\alias{two_class_final}
\title{Test set results for logistic regression}
\value{
An object with primary class \code{last_fit}.
}
\description{
This object has the results when a logistic regression model is fit to the
training set and is evaluated on the test set.
}
\details{
The code used to produce this object:\preformatted{  library(tidymodels)
  tidymodels_prefer()

  # ------------------------------------------------------------------------------

  set.seed(1)
  data(two_class_dat)

  # ------------------------------------------------------------------------------

  two_class_split <- initial_split(two_class_dat)
  # ------------------------------------------------------------------------------

  glm_spec <- logistic_reg()

  two_class_final <-
    glm_spec \%>\%
    last_fit(
      Class  ~ .,
      split = two_class_split
    )
}
}
\keyword{datasets}
