% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{display_selected}
\alias{display_selected}
\title{Gets the config and translate to a sentence with the parameter values}
\usage{
display_selected(x, predictions, tuning_param, input)
}
\arguments{
\item{x}{The \code{\link[=organize_data]{organize_data()}} result.}

\item{predictions}{The dataframe with predictions across all models.}

\item{tuning_param}{The names of the tuning parameters.}

\item{input}{The datatable object.}
}
\value{
A sentence.
}
\description{
This function takes result of \link{organize_data}, predictions across all models,
and the names of the tuning parameters to return a sentence with the default
parameter values.
}
