% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{first_class_prob_name}
\alias{first_class_prob_name}
\title{Returns the name of predictions column for the first level variable}
\usage{
first_class_prob_name(dat, event_level, y_name)
}
\arguments{
\item{dat}{The data frame obtained by merging the results from tuning functions
with the original data}

\item{event_level}{A single character value for the level corresponding to the event.}

\item{y_name}{The response variable for the model.}
}
\value{
A symbol.
}
\description{
This function takes prediction data, the event level, and the outcome name as
arguments and returns the predictions column for the first level variable.
}
\keyword{classes,}
\keyword{classif}
\keyword{models,}
