% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{first_class_prob_name}
\alias{first_class_prob_name}
\title{Returns the name of predictions column for the first level variable}
\usage{
first_class_prob_name(dat, event_level, y_name)
}
\arguments{
\item{dat}{The predictions data frame in the \code{\link[=organize_data]{organize_data()}}
result. Following variables are required: \code{.outcome}, \code{.pred}, \code{.color}, and \code{.hover}.}

\item{event_level}{A single character value for the level corresponding to the event.}

\item{y_name}{The y/response variable for the model.}
}
\value{
A symbol.
}
\description{
This function takes prediction data, the event level, and the outcome name as
arguments and returns the predictions column for the first level variable.
}
\keyword{classes,}
\keyword{classif}
\keyword{models,}
