% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{first_level}
\alias{first_level}
\title{Returns the first level of a classification model}
\usage{
first_level(dat, event_level = c("first", "second"), y_name)
}
\arguments{
\item{dat}{The predictions data frame in the \code{\link[=organize_data]{organize_data()}} result. Following
variables are required: \code{.outcome}, \code{.pred}, \code{.color}, and \code{.hover}.}

\item{event_level}{A single character value for the level corresponding to the event.}

\item{y_name}{The y/response variable for the model.}
}
\value{
A string.
}
\description{
This function takes data, \code{event_level} and \code{y_name}, as arguments and
returns the first level in a classification data.
}
\keyword{classes,}
\keyword{classif}
\keyword{models,}
