% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{first_level}
\alias{first_level}
\title{Returns the first level of a classification model}
\usage{
first_level(dat, event_level = c("first", "second"), y_name)
}
\arguments{
\item{dat}{The data frame obtained by merging the results from tuning functions
with the original data}

\item{event_level}{A single character value for the level corresponding to the event.}

\item{y_name}{The response variable for the model.}
}
\value{
A string.
}
\description{
This function takes data, \code{event_level}, and \code{y_name} as arguments and
returns the first level in a classification data.
}
\keyword{classes,}
\keyword{classif}
\keyword{models,}
