% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{first_level}
\alias{first_level}
\title{Rwturns the first level of a classification model}
\usage{
first_level(data, event_level, y_name)
}
\arguments{
\item{data}{give the dataset you are trying find the first level for}

\item{event_level}{give the original event level for the dataset}

\item{y_name}{give the response variable for the dataset}
}
\value{
first_level()
}
\description{
This function takes the data, event_level, and y_name as arguments and
returns the first level in a classification data.
}
\keyword{classes,}
\keyword{classif}
\keyword{models,}
