% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{format_hover}
\alias{format_hover}
\title{Returns the hover columns to be displayed in interactive plots}
\usage{
format_hover(x, ...)
}
\arguments{
\item{x}{.hover column from a \code{\link[=organize_data]{organize_data()}} result}

\item{...}{Other parameters not currently used.}
}
\value{
A string.
}
\description{
This function takes .hover argument and returns the output that can
be used as a test aesthetics in a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object to customize tooltip.
}
\keyword{models}
