% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Internal_data}
\alias{Internal_data}
\alias{ad_data_boost_res}
\alias{mtcars_spline_res}
\alias{ames_rf_res}
\alias{volcano_res}
\alias{svm_res}
\title{A tune_ function result object for internal testing}
\value{
a tune_() result object
}
\description{
A tune_ function result object for internal testing
}
\details{
An internal object used for development containing the
results of tune_grid(), tune_bayes(), fit_resamples(), or last_fit().
}
\keyword{datasets}
