% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Internal_data}
\alias{Internal_data}
\alias{ad_data_boost_res}
\alias{mtcars_spline_res}
\alias{ames_rf_res}
\alias{volcano_res}
\alias{svm_res}
\title{A \code{tune} package result object for internal testing}
\value{
A tune_() result object
}
\description{
A \code{tune} package result object for internal testing
}
\details{
An internal object used for development containing the
results of \code{\link[tune:tune_grid]{tune::tune_grid()}}, \code{\link[tune:tune_bayes]{tune::tune_bayes()}}, \code{\link[tune:fit_resamples]{tune::fit_resamples()}},
or \code{\link[tune:last_fit]{tune::last_fit()}}.
}
\keyword{datasets}
