% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R, R/organize_data.R
\name{organize_data}
\alias{organize_data}
\alias{organize_data.default}
\alias{organize_data.tune_results}
\title{explore()
Given an object of class \code{shiny_data}, lanches a shiny app}
\usage{
organize_data(x, ...)

\method{organize_data}{default}(x, ...)

\method{organize_data}{tune_results}(x, ...)

organize_data(x, ...)

\method{organize_data}{default}(x, ...)

\method{organize_data}{tune_results}(x, ...)
}
\arguments{
\item{x}{The \code{\link[tune:fit_resamples]{tune::fit_resamples()}} result.}

\item{...}{Other parameters not currently used.}
}
\value{
A data frame.

A data frame.
}
\description{
This function joins the result of \code{\link[tune:fit_resamples]{tune::fit_resamples()}}  to the original dataset
to give a dataframe that can be a Shiny input.

This function joins the result of \code{\link[tune:fit_resamples]{tune::fit_resamples()}}  to the original dataset
to give a dataframe that can be a Shiny input.
}
\keyword{classes,}
\keyword{classif}
\keyword{graphs,}
\keyword{models,}
\keyword{regression,}
