% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_data.R
\name{organize_data}
\alias{organize_data}
\alias{organize_data.default}
\alias{organize_data.tune_results}
\title{Extract data from objects to use in a shiny app}
\usage{
organize_data(x, cols = NULL, ...)

\method{organize_data}{default}(x, cols = NULL, ...)

\method{organize_data}{tune_results}(x, cols = NULL, ...)
}
\arguments{
\item{x}{The \code{\link[tune:fit_resamples]{tune::fit_resamples()}} result.}

\item{cols}{The columns to display while hovering.}

\item{...}{Other parameters not currently used.}
}
\value{
A data frame.
}
\description{
This function joins the result of \code{\link[tune:fit_resamples]{tune::fit_resamples()}}  to the original
dataset to give a dataframe that can be an input for a Shiny app.
}
\keyword{classes,}
\keyword{classif}
\keyword{graphs,}
\keyword{models,}
\keyword{regression,}
