% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{organize_data}
\alias{organize_data}
\title{Extracts data from objects to use in a shiny app.}
\usage{
organize_data(res_object, original_data)
}
\arguments{
\item{res_object}{The \code{\link[tune:fit_resamples]{tune::fit_resamples()}} result.}

\item{original_data}{The original dataset.}
}
\value{
A data frame.
}
\description{
This function joins the result of \code{\link[tune:fit_resamples]{tune::fit_resamples()}}  to the original dataset
to give a dataframe that can be a Shiny input.
}
\keyword{classes,}
\keyword{classif}
\keyword{graphs,}
\keyword{models,}
\keyword{regression,}
