% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{performance_object}
\alias{performance_object}
\title{This function takes result of \link{organize_data} to calculate and reformat
performance metrics for each candidate model.}
\usage{
performance_object(x, estimate_name)
}
\arguments{
\item{x}{The \code{\link[=organize_data]{organize_data()}} result.}

\item{estimate_name}{The column name with performance metrics.}
}
\value{
A dataframe.
}
\description{
This function takes result of \link{organize_data} to calculate and reformat
performance metrics for each candidate model.
}
