% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{performance_object}
\alias{performance_object}
\title{This function takes result of \code{\link[=organize_data]{organize_data()}} to calculate and reformat
performance metrics for each candidate model.}
\usage{
performance_object(x)
}
\arguments{
\item{x}{The \code{\link[=organize_data]{organize_data()}} result.}
}
\value{
A data frame.
}
\description{
This function takes result of \code{\link[=organize_data]{organize_data()}} to calculate and reformat
performance metrics for each candidate model.
}
\details{
For objects that were produced by one of the racing functions, the results
are restricted to tuning parameter combinations that have results for all of
the resamples. In other words, only the models left at the end of the race
will be shown to the users.
}
