% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diag_plots.R
\name{plot_numeric_obs_pred}
\alias{plot_numeric_obs_pred}
\title{Visualizing observed vs. predicted values for a regression model}
\usage{
plot_numeric_obs_pred(original_data, object)
}
\arguments{
\item{original_data}{give the original dataset you used for modeling}

\item{object}{give the object you got using tidymodels::fit_resamples}
}
\value{
plot_numeric_obs_pred()
}
\description{
This function allows you to plot the predicted values based on your tidymodels
results for a regression model against the observed/true values.
}
\keyword{graphs}
\keyword{models,}
\keyword{regression,}
