% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diag_plots.R
\name{plot_numeric_obs_pred}
\alias{plot_numeric_obs_pred}
\title{Visualizing observed vs. predicted values for a regression model}
\usage{
plot_numeric_obs_pred(dat, y_name)
}
\arguments{
\item{dat}{give the data frame that has the original data and the results of
collect_predictions()}

\item{y_name}{give the y/response variable for the model}
}
\value{
plot_numeric_obs_pred()
}
\description{
This function allows you to plot the predicted values based on your tidymodels
results for a regression model against the observed/true values.
}
\keyword{graphs}
\keyword{models,}
\keyword{regression,}
