% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diag_plots.R
\name{plot_numeric_res_faccol}
\alias{plot_numeric_res_faccol}
\title{Visualizing residuals vs. a factor column for a model}
\usage{
plot_numeric_res_faccol(original_data, object, factorcol)
}
\arguments{
\item{object}{give the object you got using tidymodels::fit_resamples}

\item{factorcol}{give the factor column you want to plot against the residuals}
}
\value{
plot_numeric_res_factorcol()
}
\description{
This function allows you to plot the residuals based on tidymodels result
against any of the factor column used for modeling.
}
\keyword{graphs}
\keyword{models,}
\keyword{regression,}
