% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diag_plots.R
\name{plot_numeric_res_factorcol}
\alias{plot_numeric_res_factorcol}
\title{Visualizing residuals vs. a factor column for a regression model}
\usage{
plot_numeric_res_factorcol(dat, factorcol, y_name)
}
\arguments{
\item{dat}{The data frame that has the original data and the results of
\code{\link[tune:collect_predictions]{tune::collect_predictions()}}.}

\item{factorcol}{The factor column you want to plot against the residuals.}

\item{y_name}{The y/response variable for the model.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
This function plots the residuals based on tidymodels result for
a regression model against any of the factor column used for modeling.
}
\keyword{graphs}
\keyword{models,}
\keyword{regression,}
