% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diag_plots.R
\name{plot_numeric_res_numcol}
\alias{plot_numeric_res_numcol}
\title{Visualizing residuals vs. a numeric column for a regression model}
\usage{
plot_numeric_res_numcol(dat, numcol, y_name)
}
\arguments{
\item{dat}{give the data frame that has the original data and the results of
collect_predictions()}

\item{numcol}{give the numerical column you want to plot against the residuals}

\item{y_name}{give the y/response variable for the model}
}
\value{
plot_numeric_res_numcol()
}
\description{
This function allows you to plot the residuals based on tidymodels result for
a regression model against any of the numeric column used for modeling.
}
\keyword{graphs}
\keyword{models,}
\keyword{regression,}
