% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diag_plots.R
\name{plot_numeric_res_numcol}
\alias{plot_numeric_res_numcol}
\title{Visualizing residuals vs. a numeric column for a regression model}
\usage{
plot_numeric_res_numcol(dat, y_name, numcol, alpha = 1, size = 1)
}
\arguments{
\item{dat}{The predictions data frame in the \code{\link[=organize_data]{organize_data()}} result. Following
variables are required: \code{.outcome}, \code{.pred}, \code{.color}, and \code{.hover}.}

\item{y_name}{The y/response variable for the model.}

\item{numcol}{The numerical column to plot against the residuals.}

\item{alpha}{The opacity for the geom points.}

\item{size}{The size for the geom points.}
}
\value{
A \code{\link[plotly:ggplotly]{plotly::ggplotly()}} object.
}
\description{
This function plots the residuals against a numeric column based on
{tidymodels} results for a regression model.
}
\keyword{graphs}
\keyword{models,}
\keyword{regression,}
