% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diag_plots.R
\name{plot_numeric_res_numcol}
\alias{plot_numeric_res_numcol}
\title{Residuals vs. numeric column function}
\usage{
plot_numeric_res_numcol(original_data, fit_resamples_obj, num_column)
}
\arguments{
\item{original_data}{give the original dataset you used for modeling}

\item{num_column}{give the numerical column you want to plot against the residuals}

\item{fit_samples_obj}{give the object you got using tidymodels::fit_resamples}
}
\description{
This function allows you to plot the residuals from your model against any of the numeric column used for modeling.
}
\examples{
plot_numeric_res_numcol()
}
\keyword{column}
\keyword{numerical}
\keyword{residuals}
\keyword{vs.}
