% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diag_plots.R
\name{plot_numeric_res_pred}
\alias{plot_numeric_res_pred}
\title{Visualizing residuals vs. predicted values for a regression model}
\usage{
plot_numeric_res_pred(dat, y_name)
}
\arguments{
\item{dat}{The data frame that has the original data and the results of
\code{\link[tune:collect_predictions]{tune::collect_predictions()}}.}

\item{y_name}{The y/response variable for the model.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
This function plots the predicted values based on your tidymodels
result for a regression model against the residuals.
}
\keyword{graphs}
\keyword{models,}
\keyword{regression,}
