% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diag_plots.R
\name{plot_numeric_res_pred}
\alias{plot_numeric_res_pred}
\title{Visualizing residuals vs. predicted values for a regression model}
\usage{
plot_numeric_res_pred(dat, y_name)
}
\arguments{
\item{dat}{The predictions data frame in the \code{\link[=organize_data]{organize_data()}}
result. Following variables are required: \code{.outcome}, \code{.pred}, \code{.color}, and \code{.hover}.}

\item{y_name}{The y/response variable for the model.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
This function plots the predicted values against the residuals based on
{tidymodels} results for a regression model.
}
\keyword{graphs}
\keyword{models,}
\keyword{regression,}
