% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_autoplot.R
\name{plot_tuning_params}
\alias{plot_tuning_params}
\title{Plot tuning parameters}
\usage{
plot_tuning_params(obj, source, ...)
}
\arguments{
\item{obj}{An object with class \code{tune_results}.}

\item{source}{A character string of length 1 that matches the source argument
in event_data().}

\item{...}{Other options to pass to \code{\link[tune:autoplot.tune_results]{tune::autoplot.tune_results()}}.}
}
\value{
A \code{\link[plotly:ggplotly]{plotly::ggplotly()}} object.
}
\description{
This function plots the tuning parameters against the metrics measured
during tuning.
}
\details{
The ggplot object is modified to have an aesthetic of
\code{customdata = .config}.
}
\keyword{internal}
