% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_diag_plots.R
\name{plot_twoclass_conf_mat}
\alias{plot_twoclass_conf_mat}
\title{Visualizing the confusion matrix for a classification model}
\usage{
plot_twoclass_conf_mat(object)
}
\arguments{
\item{object}{The object you got using \code{\link[tune:fit_resamples]{tune::fit_resamples()}}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
This function plots the confusion matrix for your classification
model.
}
\keyword{classes,}
\keyword{classif,}
\keyword{graphs}
\keyword{models,}
