% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_diag_plots.R
\name{plot_twoclass_obs_pred}
\alias{plot_twoclass_obs_pred}
\title{Visualizing predicted probability vs. true class for a two-class classification
model}
\usage{
plot_twoclass_obs_pred(object, event_level = "first", prob_bins = 0.025)
}
\arguments{
\item{object}{The object you got using \code{\link[tune:fit_resamples]{tune::fit_resamples()}}.}

\item{event_level}{The desired event level for the dataset.}

\item{prob_bins}{The desired \code{binwidth} for histogram.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
This function plots the predicted probabilities from your tidymodels
result for a two-class classification model against the observed/true class.
}
\keyword{classes,}
\keyword{classif,}
\keyword{graphs}
\keyword{models,}
