% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_diag_plots.R
\name{plot_twoclass_obs_pred}
\alias{plot_twoclass_obs_pred}
\title{Visualizing predicted probability vs. true class for a two-class classification
model}
\usage{
plot_twoclass_obs_pred(dat, y_name, event_level = "first", prob_bins = 0.025)
}
\arguments{
\item{dat}{The predictions data frame in the \code{\link[=organize_data]{organize_data()}}
result. Following variables are required: \code{.outcome}, \code{.pred}, \code{.color}, and \code{.hover}.}

\item{y_name}{The y/response variable for the model.}

\item{event_level}{A single character value for the level corresponding to the event.}

\item{prob_bins}{The desired \code{binwidth} for histogram.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
This function plots the predicted probabilities against the observed class based on
{tidymodels} results for a two-class classification model.
}
\keyword{classes,}
\keyword{classif,}
\keyword{graphs}
\keyword{models,}
