% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_diag_plots.R
\name{plot_twoclass_pr}
\alias{plot_twoclass_pr}
\title{Visualizing the pr curve for a classification model}
\usage{
plot_twoclass_pr(object, event_level = "first")
}
\arguments{
\item{object}{give the object you got using tidymodels::fit_resamples}

\item{event_level}{give the desired event level for the dataset}
}
\value{
plot_twoclass_pr()
}
\description{
This function allows you to plot the full precision recall curve.
}
\keyword{classes,}
\keyword{classif,}
\keyword{graphs}
\keyword{models,}
