% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_diag_plots.R
\name{plot_twoclass_pred_factorcol}
\alias{plot_twoclass_pred_factorcol}
\title{Visualizing the predicted probability vs. a factor variable for a classification
model}
\usage{
plot_twoclass_pred_factorcol(
  dat,
  factorcol,
  y_name,
  event_level = "first",
  prob_breaks = (2:9)/10,
  prob_eps = 0.001
)
}
\arguments{
\item{dat}{The data frame that has the original data and the results of
\code{\link[tune:collect_predictions]{tune::collect_predictions()}}.}

\item{factorcol}{The factor column you want to plot.}

\item{y_name}{The y/response variable for the model.}

\item{event_level}{The desired event level for the dataset.}

\item{prob_breaks}{The breaks for \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}.}

\item{prob_eps}{The threshold to break prob_name.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
This function plots the predicted probabilities from your
tidymodels result for a two-class classification model against a factor variable.
}
\keyword{classes,}
\keyword{classif,}
\keyword{graphs}
\keyword{models,}
