% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_diag_plots.R
\name{plot_twoclass_pred_factorcol}
\alias{plot_twoclass_pred_factorcol}
\title{Visualizing the predicted probability vs. a factor variable for a classification
model}
\usage{
plot_twoclass_pred_factorcol(
  dat,
  factorcol,
  y_name,
  event_level = "first",
  prob_breaks = (2:9)/10,
  prob_eps = 0.001
)
}
\arguments{
\item{dat}{give the dataframe obtained by merging the results from tuning functions
with the original data}

\item{factorcol}{give the factor column you want to plot}

\item{y_name}{give the y/response variable for the model}
}
\value{
plot_twoclass_pred_factorcol()
}
\description{
This function allows you to plot the predicted probabilities from your
tidymodels result for a two-class classification model against a factor variable.
}
\keyword{classes,}
\keyword{classif,}
\keyword{graphs}
\keyword{models,}
