% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_diag_plots.R
\name{plot_twoclass_pred_numcol}
\alias{plot_twoclass_pred_numcol}
\title{Visualizing the predicted probabilities vs. a numeric column for a
classification model}
\usage{
plot_twoclass_pred_numcol(
  dat,
  y_name,
  numcol,
  event_level = "first",
  prob_breaks = (2:9)/10,
  prob_eps = 0.001
)
}
\arguments{
\item{dat}{The predictions data frame in the \code{\link[=organize_data]{organize_data()}} result. Following
variables are required: \code{.outcome}, \code{.pred}, \code{.color}, and \code{.hover}.}

\item{y_name}{The y/response variable for the model.}

\item{numcol}{The numerical column to plot against the predicted probabilities.}

\item{event_level}{A single character value for the level corresponding to the event.}

\item{prob_breaks}{A vector to use for breaks in the probability levels.}

\item{prob_eps}{A small numerical constant to prevent division by zero.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
This function plots the predicted probabilities against a numeric column based
on {tidymodels} results for a two-class classification model.
}
\keyword{classes,}
\keyword{classif,}
\keyword{graphs}
\keyword{models,}
