% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_diag_plots.R
\name{plot_twoclass_roc}
\alias{plot_twoclass_roc}
\title{Visualizing the ROC curve for a classification model}
\usage{
plot_twoclass_roc(dat, y_name, event_level = "first")
}
\arguments{
\item{dat}{The predictions data frame in the \code{\link[=organize_data]{organize_data()}} result. Following
variables are required: \code{.outcome}, \code{.pred}, \code{.color}, and \code{.hover}.}

\item{y_name}{The y/response variable for the model.}

\item{event_level}{A single character value for the level corresponding to the event.}
}
\value{
A \code{\link[plotly:ggplotly]{plotly::ggplotly()}} object.
}
\description{
This function plots the ROC curve for a classification model.
}
\keyword{classes,}
\keyword{classif,}
\keyword{graphs}
\keyword{models,}
