% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_data.R
\name{print.shiny_data}
\alias{print.shiny_data}
\title{Returns the class, app type, y name, and number of rows of an object of
\code{shiny_data} class}
\usage{
\method{print}{shiny_data}(x, cols = c(mpg, .pred), ...)
}
\arguments{
\item{x}{an object of class shiny_data}

\item{...}{Other parameters not currently used}
}
\description{
This is a print method for a shiny_data class
}
