% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymodels-package.R
\docType{package}
\name{shinymodels-package}
\alias{shinymodels}
\alias{shinymodels-package}
\title{shinymodels: Interactive Assessments of Models}
\description{
A package that launches a 'shiny' application for 'tidymodels' results. For classification or regression models, the app can be used to determine if there is lack of fit or poorly predicted points.
}
\author{
\strong{Maintainer}: Max Kuhn \email{max@rstudio.com} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Authors:
\itemize{
  \item Shisham Adhikari \email{shisham.adhikari@rstudio.com}
  \item Julia Silge \email{julia.silge@rstudio.com} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID})
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
