% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymodels-package.R
\docType{package}
\name{shinymodels-package}
\alias{shinymodels}
\alias{shinymodels-package}
\title{shinymodels: shinymodels}
\description{
A package that launches a Shiny app for {tidymodels} results.
}
\author{
\strong{Maintainer}: Shisham Adhikari \email{shisham.adhikari@rstudio.com}

Authors:
\itemize{
  \item Max Kuhn \email{max@rstudio.com} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})
  \item Julia Silge \email{julia.silge@rstudio.com} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID})
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
